# services/web/server/src/simcore_service_webserver/api/v0/openapi.yaml
openapi: 3.0.2
info:
  title: simcore-service-webserver
  description: Main service with an interface (http-API & websockets) to the web front-end
  version: 0.38.0
servers:
- url: ''
  description: webserver
- url: http://{host}:{port}
  description: development server
  variables:
    host:
      default: localhost
    port:
      default: '8001'
paths:
...
  /v0/computations/{project_id}:
    get:
      tags:
      - computations
      - projects
      summary: Get Computation
      operationId: get_computation
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ComputationTaskGet_'
...
  /v0/computations/{project_id}:stop:
    post:
      tags:
      - computations
      - projects
      summary: Stop Computation
      operationId: stop_computation
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '204':
          description: Successful Response
...
  /v0/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: get_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
    put:
      tags:
      - projects
      summary: Replace Project
      description: Replaces (i.e. full update) a project resource
      operationId: replace_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectReplace'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '204':
          description: Successful Response
    patch:
      tags:
      - projects
      summary: Update Project
      description: Partial update of a project resource
      operationId: update_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
  /v0/projects/{project_id}:clone:
    post:
      tags:
      - projects
      summary: Clone Project
      operationId: clone_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TaskGet_'
...
  /v0/projects/{project_id}/nodes/{node_id}/outputs:
    patch:
      tags:
      - projects
      - nodes
      summary: Update Node Outputs
      operationId: update_node_outputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
        name: project_id
        in: path
      - required: true
        schema:
          title: Node Id
          type: string
        name: node_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeOutputs'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
  /v0/projects/{project_id}/inputs:
    get:
      tags:
      - projects
      - ports
      summary: Get Project Inputs
      description: New in version *0.10*
      operationId: get_project_inputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectInputGet__'
    patch:
      tags:
      - projects
      - ports
      summary: Update Project Inputs
      description: New in version *0.10*
      operationId: update_project_inputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              title: ' Updates'
              type: array
              items:
                $ref: '#/components/schemas/ProjectInputUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectInputGet__'
  /v0/projects/{project_id}/outputs:
    get:
      tags:
      - projects
      - ports
      summary: Get Project Outputs
      description: New in version *0.10*
      operationId: get_project_outputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectOutputGet__'
...
components:
  schemas:
...
    AccessRights:
      title: AccessRights
      required:
      - read
      - write
      - delete
      type: object
      properties:
        read:
          title: Read
          type: boolean
          description: has read access
        write:
          title: Write
          type: boolean
          description: has write access
        delete:
          title: Delete
          type: boolean
          description: has deletion rights
      additionalProperties: false
...
    ComputationTaskGet:
      title: ComputationTaskGet
      type: object
      properties:
        cluster_id:
          title: Cluster Id
          minimum: 0
          type: integer
...
    EmptyModel:
      title: EmptyModel
      type: object
      properties: {}
      additionalProperties: false
...
    Envelope_ComputationTaskGet_:
      title: Envelope[ComputationTaskGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ComputationTaskGet'
        error:
          title: Error
...
    Envelope_ProjectGet_:
      title: Envelope[ProjectGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ProjectGet'
        error:
          title: Error
...
    Envelope_TaskGet_:
      title: Envelope[TaskGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskGet'
        error:
          title: Error
...
    Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectInputGet__:
      title: Envelope[dict[uuid.UUID, simcore_service_webserver.projects._ports_handlers.ProjectInputGet]]
      type: object
      properties:
        data:
          title: Data
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProjectInputGet'
        error:
          title: Error
    Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectOutputGet__:
      title: Envelope[dict[uuid.UUID, simcore_service_webserver.projects._ports_handlers.ProjectOutputGet]]
      type: object
      properties:
        data:
          title: Data
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProjectOutputGet'
        error:
          title: Error
...
    NodeOutputs:
      title: NodeOutputs
      required:
      - outputs
      type: object
      properties:
        outputs:
          title: Outputs
          type: object
...
    Owner:
      title: Owner
      required:
      - user_id
      - first_name
      - last_name
      type: object
      properties:
        user_id:
          title: User Id
          type: integer
          description: Owner's user id
        first_name:
          title: First Name
          maxLength: 255
          type: string
          description: Owner's first name
        last_name:
          title: Last Name
          maxLength: 255
          type: string
          description: Owner's last name
      additionalProperties: false
...
    ProjectGet:
      title: ProjectGet
      required:
      - uuid
      - name
      - description
      - thumbnail
      - creationDate
      - lastChangeDate
      - workbench
      - prjOwner
      - accessRights
      - tags
      type: object
      properties:
        uuid:
          title: Uuid
          type: string
          format: uuid
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        thumbnail:
          title: Thumbnail
          anyOf:
          - maxLength: 2083
            minLength: 0
            type: string
            format: uri
          - enum:
            - ''
            type: string
        creationDate:
          title: Creationdate
          pattern: \d{4}-(12|11|10|0?[1-9])-(31|30|[0-2]?\d)T(2[0-3]|1\d|0?[0-9])(:(\d|[0-5]\d)){2}(\.\d{3})?Z
          type: string
        lastChangeDate:
          title: Lastchangedate
          pattern: \d{4}-(12|11|10|0?[1-9])-(31|30|[0-2]?\d)T(2[0-3]|1\d|0?[0-9])(:(\d|[0-5]\d)){2}(\.\d{3})?Z
          type: string
        workbench:
          title: Workbench
          type: object
        prjOwner:
          title: Prjowner
          type: string
          format: email
        accessRights:
          title: Accessrights
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AccessRights'
        tags:
          title: Tags
          type: array
          items:
            type: integer
        classifiers:
          title: Classifiers
          type: array
          items:
            type: string
          default: []
        state:
          $ref: '#/components/schemas/ProjectState'
        ui:
          title: Ui
          anyOf:
          - $ref: '#/components/schemas/EmptyModel'
          - $ref: '#/components/schemas/StudyUI'
        quality:
          title: Quality
          type: object
          default: {}
        dev:
          title: Dev
          type: object
        permalink:
          $ref: '#/components/schemas/ProjectPermalink'
    ProjectInputGet:
      title: ProjectInputGet
      required:
      - key
      - value
      - label
      type: object
      properties:
        key:
          title: Key
          type: string
          description: Project port's unique identifer. Same as the UUID of the associated
            port node
          format: uuid
        value:
          title: Value
          description: Value assigned to this i/o port
        label:
          title: Label
          type: string
    ProjectInputUpdate:
      title: ProjectInputUpdate
      required:
      - key
      - value
      type: object
      properties:
        key:
          title: Key
          type: string
          description: Project port's unique identifer. Same as the UUID of the associated
            port node
          format: uuid
        value:
          title: Value
          description: Value assigned to this i/o port
...
    ProjectLocked:
      title: ProjectLocked
      required:
      - value
      - status
      type: object
      properties:
        value:
          title: Value
          type: boolean
          description: True if the project is locked
        owner:
          title: Owner
          allOf:
          - $ref: '#/components/schemas/Owner'
          description: If locked, the user that owns the lock
        status:
          allOf:
          - $ref: '#/components/schemas/ProjectStatus'
          description: The status of the project
      additionalProperties: false
...
    ProjectOutputGet:
      title: ProjectOutputGet
      required:
      - key
      - value
      - label
      type: object
      properties:
        key:
          title: Key
          type: string
          description: Project port's unique identifer. Same as the UUID of the associated
            port node
          format: uuid
        value:
          title: Value
          description: Value assigned to this i/o port
        label:
          title: Label
          type: string
    ProjectPermalink:
      title: ProjectPermalink
      required:
      - url
      - is_public
      type: object
      properties:
        url:
          title: Url
          maxLength: 2083
          minLength: 1
          type: string
          format: uri
        is_public:
          title: Is Public
          type: boolean
    ProjectReplace:
      title: ProjectReplace
      required:
      - uuid
      - name
      - description
      - creationDate
      - lastChangeDate
      - workbench
      - accessRights
      type: object
      properties:
        uuid:
          title: Uuid
          type: string
          format: uuid
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        thumbnail:
          title: Thumbnail
          maxLength: 2083
          minLength: 0
          type: string
          format: uri
        creationDate:
          title: Creationdate
          pattern: \d{4}-(12|11|10|0?[1-9])-(31|30|[0-2]?\d)T(2[0-3]|1\d|0?[0-9])(:(\d|[0-5]\d)){2}(\.\d{3})?Z
          type: string
        lastChangeDate:
          title: Lastchangedate
          pattern: \d{4}-(12|11|10|0?[1-9])-(31|30|[0-2]?\d)T(2[0-3]|1\d|0?[0-9])(:(\d|[0-5]\d)){2}(\.\d{3})?Z
          type: string
        workbench:
          title: Workbench
          type: object
        accessRights:
          title: Accessrights
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AccessRights'
        tags:
          title: Tags
          type: array
          items:
            type: integer
          default: []
        classifiers:
          title: Classifiers
          type: array
          items:
            type: string
        ui:
          $ref: '#/components/schemas/StudyUI'
        quality:
          title: Quality
          type: object
    ProjectRunningState:
      title: ProjectRunningState
      required:
      - value
      type: object
      properties:
        value:
          allOf:
          - $ref: '#/components/schemas/RunningState'
          description: The running state of the project
      additionalProperties: false
    ProjectState:
      title: ProjectState
      required:
      - locked
      - state
      type: object
      properties:
        locked:
          title: Locked
          allOf:
          - $ref: '#/components/schemas/ProjectLocked'
          description: The project lock state
        state:
          title: State
          allOf:
          - $ref: '#/components/schemas/ProjectRunningState'
          description: The project running state
      additionalProperties: false
    ProjectStatus:
      title: ProjectStatus
      enum:
      - CLOSED
      - CLOSING
      - CLONING
      - EXPORTING
      - OPENING
      - OPENED
      type: string
      description: An enumeration.
...
    ProjectUpdate:
      title: ProjectUpdate
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        thumbnail:
          title: Thumbnail
          maxLength: 2083
          minLength: 0
          type: string
          format: uri
        workbench:
          title: Workbench
          type: object
        accessRights:
          title: Accessrights
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AccessRights'
        tags:
          title: Tags
          type: array
          items:
            type: integer
        classifiers:
          title: Classifiers
          type: array
          items:
            type: string
        ui:
          $ref: '#/components/schemas/StudyUI'
        quality:
          title: Quality
          type: object
...
    RunningState:
      title: RunningState
      enum:
      - UNKNOWN
      - PUBLISHED
      - NOT_STARTED
      - PENDING
      - WAITING_FOR_RESOURCES
      - STARTED
      - SUCCESS
      - FAILED
      - ABORTED
      - WAITING_FOR_CLUSTER
      type: string
      description: 'State of execution of a project''s computational workflow


        SEE StateType for task state'
...
    StudyUI:
      title: StudyUI
      type: object
      properties:
        workbench:
          title: Workbench
          type: object
        slideshow:
          title: Slideshow
          type: object
        currentNodeId:
          title: Currentnodeid
          type: string
          format: uuid
        annotations:
          title: Annotations
          type: object
...
    TaskGet:
      title: TaskGet
      required:
      - task_id
      - task_name
      - status_href
      - result_href
      - abort_href
      type: object
      properties:
        task_id:
          title: Task Id
          type: string
        task_name:
          title: Task Name
          type: string
        status_href:
          title: Status Href
          type: string
        result_href:
          title: Result Href
          type: string
        abort_href:
          title: Abort Href
          type: string
...